put 'active' tab indication back.
put styling for active tab in <head> because
setting an .active class in the partial wont work
since it is cached.
And it can't be done in css with all combinations of:

body.whatever #section-tabs .whatever,
body.otherthing #section-tabs .otherthing {…}

because by adding a BLTI tab, you would not know
the class until runtime.

test plan:
in an environment with caching turned on,
verify that it looks like this:
http://cl.ly/2T0N0f0h1B0S1K070U0D
not this:
http://cl.ly/0e0T430J1Y12211C2r29

fixes: #6791

Change-Id: I10d405f7fdf8d8edce8cec3e74ad62ec977a4676
Reviewed-on: https://gerrit.instructure.com/7887
Tested-by: Hudson <hudson@instructure.com>
Reviewed-by: Zach Wily <zach@instructure.com>