don't store broadcast policy procs on AR objects
After saving an object that has a broadcast policy, it has a bunch of
procs stored in an instance variable. This breaks Marshal, and thus
caching.

This doesn't fix any known bugs, it's just a precaution against future
issues arising.

refs #6867

test plan: This isn't reproducable from the app, though it's
simple enough from the rails console.

Change-Id: If47b4618966ac82d472fc8d2303fd333aa2001ce
Reviewed-on: https://gerrit.instructure.com/8049
Reviewed-by: Cody Cutrer <cody@instructure.com>
Tested-by: Hudson <hudson@instructure.com>