don't fail when copying/importing large courses
Rack has a set limit for how large the "keyspace" for
requests can be. For courses with lots of items the request
exceeded this limit. This commit changes the copy/import
forms to have a single list of items to import instead of
lots of key/value pairs.

Test Plan:
 * Manually run some course copies and content imports

closes #6889

Change-Id: I18b1e49261ef8c4ea724a8e0d02ac6bbdaac2fbc
Reviewed-on: https://gerrit.instructure.com/8120
Reviewed-by: Brian Palmer <brianp@instructure.com>
Tested-by: Hudson <hudson@instructure.com>