fix zendesk not getting the right user on new tickets
In the old form, the form would post a blank
report[email] value.  because of the way
guess_email works we can't just not submit it. 
because it needs to override the "unknown" value 
that was put there in the after_save callback when
it was created (but did not yet have an associated
user)

test_plan:
 * run the spec 
Manual steps:
 * in an environment with the canvas_zendesk_plugin
   set up, submit something a problem from the 
   help dialog
 * make sure the ticket in zendesk has the right
   email address.

Change-Id: I23d983da41e154815c7f28fe64fb70f03a0a9803
fixes: #7018
Reviewed-on: https://gerrit.instructure.com/8255
Reviewed-by: Jon Jensen <jon@instructure.com>
Tested-by: Hudson <hudson@instructure.com>