allow plugins to extend canvas.rake
Rails expects plugins to not extend the app, but provide services for
the app, so it loads plugin tasks first, then the app tasks.  So for
a plugin to modify the behavior of these tasks, it has to manually
require the file.  But Rails will then re-load the file, and due to
how Rake works, these tasks will all be duplicated.  So do manual
single-loading.

test plan:
 * make sure rake db:pending_migrations still works

Change-Id: I54458a085ee1b771093734d069472a55c4309d25
Reviewed-on: https://gerrit.instructure.com/8483
Tested-by: Hudson <hudson@instructure.com>
Reviewed-by: Brian Palmer <brianp@instructure.com>