less guessable avatar urls
Avatar URLs were just based on the user id, so it was
trivial for someone to wander the id space and find valid
pictures of users. This commit adds a hash to the id so it's
no longer guessable, but still retrievable. This commit
still allows old-style urls to work, so we can give caches
time to clear out before we disabled the old style.

Also moved the avatar action to the users controller since it
makes a lot more sense there.

fixes #5618

test plan:
- make sure the old style of avatar image urls still work

- go to a page with user avatars (like discussions)
- make sure avatars appear correctly for users with avatars

- put in a bogus URL ("/images/users/a")
- make sure it doesn't die

- put in a bogus URL with a hyphen ("/images/users/1-1")
- make sure it doesn't die

Change-Id: I796dffc2b3f00a4b9db082e1fcbd289f47a62c72
Reviewed-on: https://gerrit.instructure.com/8938
Reviewed-by: Brian Palmer <brianp@instructure.com>
Tested-by: Hudson <hudson@instructure.com>