better loading indication in messages, fixes: #7377
change all of the loading indicators in messages
to the newer 'disableWhileLoading' spinner instead
of the not-as-good-looking red spinner gif.

this has the side-effect that it also makes it so
you can't submit the form twice.

so for example:

this: http://cl.ly/3m3G2C1m380l150E3r0X
instead of: http://cl.ly/1I0o1W3J0K3a2t0b2d0y

Change-Id: I43045785b08690e518dda353f5dfc636bda99a50
Reviewed-on: https://gerrit.instructure.com/8555
Tested-by: Hudson <hudson@instructure.com>
Reviewed-by: Jon Jensen <jon@instructure.com>