make question data always a HashWithIndifferentAccess
All the code that works with question data is dependant on
the data being a HashWithIndifferentAccess. This commit
makes sure to use that in AQs and QQs, and made the
accessor to question_data set it to the expected hash
type so that old incorrectly serialized data doesn't have
to be updated.

Test Plan:
 * Create some assessment questions that will run through translate_links
 * When accesssing that data id should be an indifferent access hash
 * When viewing those questions in the UI they should display correctly for students and teachers
 * Manually set question data as a hash on a AQ or QQ, when you access it through the AQ/QQ it should be a HWIA

closes #6897

Change-Id: I526ab886aad2f96a3741330059439b100b6a1011
Reviewed-on: https://gerrit.instructure.com/8623
Tested-by: Hudson <hudson@instructure.com>
Reviewed-by: Zach Wily <zach@instructure.com>