fix assigning of assignment group when no active groups exist
fixes #7275

Because whiny_nils on raises an exception in dev/test modes, the ".id
rescue nil" line was setting the id to nil in dev/test, but to 4
(nil.id) in production.

I reworked this so that we should always get an assignment group, and we
should never use assignment group id of 4.

test plan:
  * in production mode, create an assignment in a course that has no
    active assignment groups. this is easiest through the API, since the
    UI does stuff to ensure an assignment group always exists.
  * verify that a new group is created and the assignment is added to
    that group.

Change-Id: I34aace583c559fa9e584026612a97298177e281c
Reviewed-on: https://gerrit.instructure.com/8725
Reviewed-by: Bracken Mosbacker <bracken@instructure.com>
Tested-by: Hudson <hudson@instructure.com>