correctly update quiz submisison when changing grade in gradebook
Before only the Submission was being updated and not the
QuizSubmission. Now the fudge points on the current QS are
updated to reflect the score set in the gradebook.

This prevented modules from progressing if the value was
changed in the gradebook.

Because the score set there should be the absolute score,
if the quiz is set to take the highest a flag is set to
prevent a previously-higher score from being the kept_score

Test Plan:
 * Create a quiz with a question with 5 points and publish it
 * Create a module item to to that quiz that needs 10 points to complete
 * Take the quiz as a student, the module should not unlock
 * As the teacher set the grade to 11 in the gradebook
 * The module item should be unlocked for the student

closes #7331 #4174

Change-Id: I30d92bfe865584b5b526cda36ef336348a647de4
Reviewed-on: https://gerrit.instructure.com/9091
Reviewed-by: Cody Cutrer <cody@instructure.com>
Tested-by: Hudson <hudson@instructure.com>