spec for excluding unmanageable permissions in api
in the role api, only those permissions that can be managed within the
account the role belongs to should be returned in responses. e.g. the
"become_user" permission can only be managed on root accounts; when
creating a role on a subaccount they shouldn't see "become_user" in the
response.

the code was already correct, but noticed we were lacking a spec for it.

test-plan: run the specs

Change-Id: I520f06f45846ee457cc16c69c54aaa0368887a09
Reviewed-on: https://gerrit.instructure.com/9176
Tested-by: Hudson <hudson@instructure.com>
Reviewed-by: Simon Williams <simon@instructure.com>