lock discussion_topic_participant row for updating when updating read counts
We need to lock this to serialize updates of the count, so they don't get off
due to race conditions. We also have to make sure the AR query cache is
disabled when using the unique_constraint_retry. If it's not, then on the
retry we might just get the results from the first query out of the cache.

It's possible that that might be generally applicable to all
unique_constraint_retry usage, but it hasn't been thought through yet.

test plan:
 * in a production-like environment:
 * create a discussion topic, and some short entries on it (at least 3)
 * load that topic with a tall window as a lot of different users
 * verify that the visible posts get marked as read, and on reload, the
   unread count stays correct.

Change-Id: I3e0cbebe991bbaaf74f4c8408cbe2214fca80eec
Reviewed-on: https://gerrit.instructure.com/9469
Reviewed-by: Simon Williams <simon@instructure.com>
Tested-by: Hudson <hudson@instructure.com>