mod indent change from edit form keeps due-date/points; fixes #7811
there was a problem where changing the indent level of a module item caused the
due-date and points possible for that item to disappear. we previously fixed
this issue when the indent was changed with the arrows, but missed the case
where the indent changes from the edit form.

test-plan:
- create an module and add an item that has a due-date and/or points
- change the indent level by clicking on the edit icon and using the form
- the due-date and/or points possible should not disappear (they may flash off
  and back)

Change-Id: Ie6ee87e70d31a56b384440ec02d7dd0d62d56e19
Reviewed-on: https://gerrit.instructure.com/9563
Tested-by: Hudson <hudson@instructure.com>
Reviewed-by: Cameron Matheson <cameron@instructure.com>