spec: detect when ajax requests will never time out due to a js error
When there is a js error during ajax response handling, the ajaxStop event
never fires. This causes wait_for_ajax_requests to wait until the script
timeout, which appears to kill the selenium thread. This catches the timeout
in javascript before the script timeout fires, and raises an exception to give
a clue about what happened.

Change-Id: I51e0de43e548e1af5a5ca2dfd793878fb3026b9f
Reviewed-on: https://gerrit.instructure.com/9761
Tested-by: Hudson <hudson@instructure.com>
Reviewed-by: Jon Jensen <jon@instructure.com>