protect clean_return_to against malicious urls
It was possible to pass in a url crafted to send the user to another
domain, or to execute javascript

fixes #8220

test plan: this shouldn't have any effect on normal usage. verify that
if you create a new assignment and go to more options, the "return" link
should take you back to the assignments page.

Change-Id: If5af6103d8ccc61b344ecbc72e131b91628457ff
Reviewed-on: https://gerrit.instructure.com/10121
Tested-by: Hudson <hudson@instructure.com>
Reviewed-by: Jon Jensen <jon@instructure.com>