update last completed quiz sub. when manually scoring; fixes #7916
a teacher can manually override the score on a quiz, and we try to keep things
in sync by writing this score back to the quiz submission. however, the quiz
submission might currently be in progress (or the latest attempt might be
abandoned) in which case syncing back can fail or do the wrong thing. So we
needed to make sure that we synced back to the correct version.

test-plan:
- take a quiz, complete it, then start it again and leave the page without
  completing it.
- as a teacher, score this quiz, and reload the page; the score should stick
- make sure scoring an essay question quiz from the quiz submission page still
  works.

Change-Id: Id2c919199a12ea8b0636959142efc649d5daf738
Reviewed-on: https://gerrit.instructure.com/10177
Tested-by: Hudson <hudson@instructure.com>
Reviewed-by: Brian Palmer <brianp@instructure.com>