properly handle time manipulations around DST-end
when adding(/subtracting) milliseconds, seconds, minutes, or hours
to(/from) a date object, work in the UTC space. if working in the local
timezone and the arithmetic takes it across the DST-end boundary, it
will not behave properly.

in addition to affecting addMinutes and friends, it also affected
Date.parse when given a time between 1:00am daylight time and 1:59am
daylight time on the day of DST-end. Date.parse constructs the date as
1:00am standard time, then compares the date's 'native' tz to the
specified tz. noticing the 1hr difference in timezones from standard to
daylight, it subtracts the hour from the new date using addMinutes --
which was broken before this commit.

test-plan:
  - load up a page with date.js on it
  - in the js console:
      date = new Date(2012, 10, 4, 0, 59, 59)
      date.addSeconds(1)
    should return 1:00am daylight time, not 0:00am daylight time
  - in the js console:
      Date.parse('2012-11-04T01:00:00-0600')
    should return 1:00am MDT, not 0:00am MDT

Change-Id: If32c42a7c53bb2eb7e6b71dd2d4e93d8e4a385e4
Reviewed-on: https://gerrit.instructure.com/10509
Tested-by: Jenkins <jenkins@instructure.com>
Reviewed-by: Brian Palmer <brianp@instructure.com>