Remove user's avatar url when profile picture is deleted. Fixes #8225
When an file attachment is deleted, check to see if it is used as the user's
profile picture. If so, delete the avatar url that references the file.
Also adds an "active" named scope to Attachment that is used in FilesController
to not return image_thumbnails for deleted attachments.
Adds a User method "clear_avatar_image_url_with_uuid" and specs to test it.

Test Plan:
* Add a profile picture to a user.
* Verify that the image shows up on GradeBook2
* Delete the file used as the profile picture from the user's "Files" area.
* Revisit the GradeBook2 page and the user's profile image should be the
   default one used when none is assigned (silhouette).

Change-Id: Iaa0941a84503e341158615a12c62a77e951c283f
Reviewed-on: https://gerrit.instructure.com/10654
Tested-by: Jenkins <jenkins@instructure.com>
Reviewed-by: Jon Jensen <jon@instructure.com>