score unknown question types as 0 / points_possible
Pre-refactor, this is how we handled unknown question types. The
refactor inadvertently made this case raise an exception. We prefer the
previous handling, though of course even better is not having this
invalid data appear in the quiz at all.

fixes #8608

test plan: Import a QTI package with an unknown question type, or insert
an invalid question type into the quiz data. Take the quiz and verify it
gets graded and you can view the submission without error.

Change-Id: I74bab7e0061cc8bed64e5385de6c4ba4799779fe
Reviewed-on: https://gerrit.instructure.com/10733
Reviewed-by: Cody Cutrer <cody@instructure.com>
Tested-by: Jenkins <jenkins@instructure.com>